requirements.txt
# Python version
python>=3.11.0

# Core dependencies
mistralai>=1.0.0
Pillow>=10.2.0
PyPDF2>=3.0.1
google-generativeai>=0.7.0
langchain>=0.1.11
langchain-google-genai>=0.0.10
faiss-cpu>=1.8.0
python-docx>=1.1.0
streamlit>=1.32.0

# Additional utilities
numpy>=1.26.0
pandas>=2.2.0
tqdm>=4.66.0
pydantic>=2.6.0
python-dotenv>=1.0.0